Added comment to lump sums, fixed layout of HTML and PDF
In the editor there is now anadditional field for every lump sum to add
a comment

The photos as well as the signatures were often crossing page
boundaries, especially signatures were far to big. Photos are now scaled
to the page width, signatures to half the page width. Ideally we would
calc the size of the signature pad on the screen and take this as size
for the signature - not now :)